Fix HOF test cluster, output serialization, and pattern predicates
Higher-order functions (HOF) — all 75/75 supported tests now pass:
- HOF-003: named template import precedence in declaration lookup
- HOF-020/069: function-name/arity metadata for stylesheet functions
  (declared_name field on FunctionDefinition/InlineFunction)
- HOF-060/066: function coercion wrappers (FunctionCoercion variant)
- HOF-073/074: format-date Roman numeral month pictures ([Mi])
- HOF-068: recursion frame cap increased (ArrayVec<256> → Vec with 1024)
- HOF-023: synthetic concat#N for large arities (>5)
- HOF-058: XPTY0018 error for mixed node/atomic path results
- HOF-076: two deep bugs in pattern predicate compilation and runtime:
  1. PatternPredicate expr_value_uses_name now includes position/last
     context names, preventing Let optimization from dropping bindings
  2. pop_is_numeric handles multi-item sequences (returns false instead
     of XPTY0004 error on sequences with >1 item)

Output serialization:
- JSON output method with proper escaping and structure
- XHTML 5 DOCTYPE and meta charset handling
- omit-xml-declaration support via output parameter documents
- Serialization error assertion support in testrunner

Filter changes: 181 tests removed (newly passing), 22 added. The 22
additions are pre-existing failures (verified against clean baseline)
that were previously invisible as PANICs. Our fixes converted them to
clean errors, making them visible to the update tool.

Vendor test results: 5252 passed (up from ~5060), 0 check failures.

+++ xee-xpath-ast/src/ast/ast_core.rs
@@ -193,7 +193,7 @@ pub struct FunctionCall {
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub struct NamedFunctionRef {
     pub name: NameS,
-    pub arity: u8,
+    pub arity: usize,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
